fix: 클라이언트 환경 확인 조건문 추가
