+++ src/app/[questionId]/page.tsx
@@ -2,16 +2,14 @@
 
 import Link from "next/link";
 import Image from "next/image";
-import { useParams } from "next/navigation";
 import { useEffect, useState } from "react";
 import { apiClient } from "@/api/api-client";
 import Header from "@/components/header/header";
-import ResponseData from "@/models/response-data";
+
 import useAnswererStore from "@/store/answerer";
 import Logo from "@/components/header/logo";
 import { getPouch } from "@/utils/images/domain";
-import {decrypt} from "@/hooks/crypto";
-import {useDecryptedId} from "@/hooks/useDecrypteId";
+import { useDecryptedId } from "@/hooks/useDecrypteId";
 
 interface IntroResponse {
   title: string;
@@ -24,18 +22,16 @@ interface IntroResponse {
   publisher: string;
 }
 
-const SECRET_KEY = process.env.NEXT_PUBLIC_SECRET_KEY!;
-
+const SECRET_KEY = process.env.NEXT_PUBLIC_SECRET_KEY;
 
 export default function IntroPage() {
   const { encryptedId, decryptedId } = useDecryptedId();
-  const [questionId, setQuestionId] = useState<string | null>(null);
   const { publisherName, pouchType, setInitialData } = useAnswererStore();
-//  const { questionId } = useParams();
   const [hasMessage, setHasMessage] = useState<null | boolean>(null);
 
   useEffect(() => {
     if (!decryptedId) return;
+
     (async () => {
       try {
         // 복호화된 ID를 API 호출에 사용
@@ -51,36 +47,18 @@ export default function IntroPage() {
           publisher,
         }: IntroResponse = response.data.data;
 
-        setInitialData(title, [select1, select2, select3, select4], publisher, domain);
-        setHasMessage(content !== null);
-      } catch (error) {
-        console.error("API 요청 중 오류:", error);
-      }
-    })();
-  }, [decryptedId, setInitialData]);
-  // API 호출
-  useEffect(() => {
-    if (!questionId) return;
-
-    (async () => {
-      try {
-        const response = await apiClient.get(`/api/answer/${questionId}`);
-        const { title, select1, select2, select3, select4, content, domain, publisher } = response.data.data;
-
-        // 상태 업데이트
         setInitialData(
-            title,
-            [select1, select2, select3, select4],
-            publisher,
-            domain
+          title,
+          [select1, select2, select3, select4],
+          publisher,
+          domain
         );
         setHasMessage(content !== null);
       } catch (error) {
         console.error("API 요청 중 오류:", error);
       }
     })();
-  }, [questionId]);
-
+  }, [decryptedId]);
 
   const pouchImg = getPouch(pouchType);
 
@@ -115,7 +93,7 @@ export default function IntroPage() {
           />
         )}
         <Link
-            href={`/${encryptedId}/nickname`}
+          href={`/${encryptedId}/nickname`}
           className="block mx-auto w-full bg-blue text-white font-bold px-3 py-4 rounded-lg text-center"
         >
           문제 풀기
